[BERT/PyT] Verify Wiki archive download success
At present, the WikiDownloader for BERT/PyTorch expects that if the
wikicorpus archive is in the local filesystem, then it was successfully
downloaded without interrpution. If, however, the download was
interrupted for any number of reasons and the WikiDownloader is run
again, it will assume the archive is valid and attempt to extract or
pre-process a corrupt dataset. Since it appears the dataset is valid
in the local filesystem, these errors that are encountered later on are
very hard to debug as they don't indicate the error exists with the
downloaded archive.

By pulling the latest checksums file from the WikiMedia repository, the
local archive if present can be compared against the expected checksum.
If the local archive's hash and the expected checksum match, then it is
safe to assume the dataset is valid and extraction can continue.

Additionally, the `create_datasets_from_start.sh` script should
terminate if any of the included commands fail as a failed command
likely means something is wrong with the setup or connections and it
needs to be investigated before continuing with later steps that would
likely fail silently as well.

Signed-Off-By: Robert Clark <roclark@nvidia.com>